Added support for ssh-nonfree

+++ gforge/deb-specific/install-ssh.sh
@@ -15,27 +15,47 @@ else
 				$0 purge
 				;;
 			chroot)
-				/etc/init.d/ssh stop
+				[ -f /etc/init.d/ssh ] && /etc/init.d/ssh stop
+				[ -f /etc/init.d/ssh-nonfree ] && /etc/init.d/ssh-nonfree stop
 				[ -f /var/lib/sourceforge/chroot/var/run/ssh.pid ] && kill $(cat /var/lib/sourceforge/chroot/var/run/ssh.pid)
+				[ -f /var/lib/sourceforge/chroot/var/run/ssh-nonfree.pid ] && kill $(cat /var/lib/sourceforge/chroot/var/run/ssh-nonfree.pid)
 				rm -f /var/lib/sourceforge/chroot/var/run/ssh.pid
+				rm -f /var/lib/sourceforge/chroot/var/run/ssh-nonfree.pid
 			        do_config=$(grep ^do_config= /etc/sourceforge/sourceforge.conf | cut -d= -f2-)
     			        if [ ! "$do_config" = "true" ] ; then
-				    if ! grep -q "start-stop-daemon --chroot /var/lib/sourceforge/chroot" /etc/init.d/ssh ; then
-					perl -pi -e "s:start-stop-daemon:start-stop-daemon --chroot /var/lib/sourceforge/chroot:g" /etc/init.d/ssh
-				    fi
-				    rm -f /etc/ssh/sshd_not_to_be_run
+					if [ -f /etc/init.d/ssh ] ; then
+				    		if ! grep -q "start-stop-daemon --chroot /var/lib/sourceforge/chroot" /etc/init.d/ssh ; then
+							perl -pi -e "s:start-stop-daemon:start-stop-daemon --chroot /var/lib/sourceforge/chroot:g" /etc/init.d/ssh
+				    		fi
+				    		rm -f /etc/ssh/sshd_not_to_be_run
+					fi
+					if [ -f /etc/init.d/ssh-nonfree ] ; then
+				    		if ! grep -q "start-stop-daemon --chroot /var/lib/sourceforge/chroot" /etc/init.d/ssh-nonfree ; then
+							perl -pi -e "s:start-stop-daemon:start-stop-daemon --chroot /var/lib/sourceforge/chroot:g" /etc/init.d/ssh-nonfree
+				    		fi
+				    		rm -f /etc/ssh/sshd-nonfree_not_to_be_run
+					fi
 				fi
-				/etc/init.d/ssh start
-				    
+				[ -f /etc/init.d/ssh ] && /etc/init.d/ssh start
+				[ -f /etc/init.d/ssh-nonfree ] && /etc/init.d/ssh-nonfree start
 				;;
 			purge)
 			        do_config=$(grep ^do_config= /etc/sourceforge/sourceforge.conf | cut -d= -f2-)
     			        if [ ! "$do_config" = "true" ] ; then
-				    if grep -q "start-stop-daemon --chroot /var/lib/sourceforge/chroot" /etc/init.d/ssh
-					then
-					perl -pi -e "s:start-stop-daemon --chroot /var/lib/sourceforge/chroot:start-stop-daemon:g" /etc/init.d/ssh
-					/etc/init.d/ssh restart
-				    fi
+					if [ -f /etc/init.d/ssh ] ; then
+				    		if grep -q "start-stop-daemon --chroot /var/lib/sourceforge/chroot" /etc/init.d/ssh
+						then
+							perl -pi -e "s:start-stop-daemon --chroot /var/lib/sourceforge/chroot:start-stop-daemon:g" /etc/init.d/ssh
+							/etc/init.d/ssh restart
+				    		fi
+					fi
+					if [ -f /etc/init.d/ssh-nonfree ] ; then
+				    		if grep -q "start-stop-daemon --chroot /var/lib/sourceforge/chroot" /etc/init.d/ssh-nonfree
+						then
+							perl -pi -e "s:start-stop-daemon --chroot /var/lib/sourceforge/chroot:start-stop-daemon:g" /etc/init.d/ssh-nonfree
+							/etc/init.d/ssh-nonfree restart
+				    		fi
+					fi
 				fi
 				;;
 		esac
